rename files acording to the guidelines and fix storybook errors related to naming
