+++ src/client/components/Auth/AuthenticatedRoute.component.js
