+++ src/client/components/CardProfileResult/CardProfileResult.styles.css
