+++ src/client/components/PageHeader/PageHeader.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import PageHeader from './PageHeader';
+import PageHeader from './PageHeader.component';
 
 export default { title: 'PageHeader' };
