+++ src/client/components/RegisterTeaser/RegisterTeaser.styles.css
